Combines the assemble and publish workflows
The assemble and publish workflows worked with the same release files and just had different purposes. Normally I'd like to keep them separated, but the process of creating an artifact from one workflow to the next with a folder this large takes a minute or two, and frankly, I'd rather not wait.

This commit combines them into one larger assemble-and-publish workflow.